+++ backend/core/src/applications/types/form-metadata/form-metadata-options.ts
@@ -18,4 +18,12 @@ export class FormMetadataOptions {
   @Type(() => FormMetadataExtraData)
   @ApiProperty({ type: [FormMetadataExtraData], required: false })
   extraData?: FormMetadataExtraData[] | null
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  description?: boolean
+
+  @Expose()
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  exclusive?: boolean
 }