+++ sites/partners/lib/hooks.ts
@@ -13,7 +13,7 @@ export function useSingleListingData(listingId: string) {
   const { listingsService } = useContext(ApiClientContext)
   const fetcher = () => listingsService.retrieve({ listingId })
 
-  const { data, error } = useSWR(process.env.listingServiceUrl, fetcher)
+  const { data, error } = useSWR(`${process.env.backendApiBase}/listings/${listingId}`, fetcher)
 
   return {
     listingDto: data,
@@ -26,7 +26,7 @@ export function useListingsData() {
   const { listingsService } = useContext(ApiClientContext)
   const fetcher = () => listingsService.list()
 
-  const { data, error } = useSWR(process.env.listingServiceUrl, fetcher)
+  const { data, error } = useSWR(`${process.env.backendApiBase}/listings`, fetcher)
 
   return {
     listingDtos: data,