+++ sites/public/cypress/integration/pages/application/review/terms.spec.ts
@@ -21,16 +21,16 @@ describe("applications/review/terms", function () {
   })
 
   // Broken on master, addressed by PR #1155
-  // it("Should redirect to the next step", function () {
-  //   cy.getByID("agree").check()
+  it("Should redirect to the next step", function () {
+    cy.getByID("agree").check()
 
-  //   submitApplication()
+    submitApplication()
 
-  //   cy.checkErrorAlert("not.exist")
-  //   cy.checkErrorMessages("not.exist")
+    cy.checkErrorAlert("not.exist")
+    cy.checkErrorMessages("not.exist")
 
-  //   cy.location("pathname").should("include", "applications/review/confirmation")
+    cy.location("pathname").should("include", "applications/review/confirmation")
 
-  //   cy.getByID("confirmationId").should("be.visible").and("not.be.empty")
-  // })
+    cy.getByID("confirmationId").should("be.visible").and("not.be.empty")
+  })
 })