+++ sites/public/pages/applications/contact/alternate-contact-type.tsx
@@ -20,7 +20,7 @@ import { useForm } from "react-hook-form"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
 
-export default () => {
+const ApplicationAlternateContactType = () => {
   const { conductor, application, listing } = useFormConductor("alternateContactType")
   const currentPageSection = 1
 
@@ -53,7 +53,10 @@ export default () => {
         />
       </FormCard>
       <FormCard>
-        <FormBackLink url={conductor.determinePreviousUrl()} />
+        <FormBackLink
+          url={conductor.determinePreviousUrl()}
+          onClick={() => conductor.setNavigatedBack(true)}
+        />
 
         <div className="form-card__lead border-b">
           <h2 className="form-card__title is-borderless">
@@ -124,7 +127,12 @@ export default () => {
           </div>
           <div className="form-card__pager">
             <div className="form-card__pager-row primary">
-              <Button styleType={AppearanceStyleType.primary}>{t("t.next")}</Button>
+              <Button
+                styleType={AppearanceStyleType.primary}
+                onClick={() => conductor.setNavigatedBack(false)}
+              >
+                {t("t.next")}
+              </Button>
             </div>
           </div>
         </Form>
@@ -132,3 +140,5 @@ export default () => {
     </FormsLayout>
   )
 }
+
+export default ApplicationAlternateContactType