+++ sites/public/pages/applications/review/confirmation.tsx
@@ -16,7 +16,7 @@ import FormsLayout from "../../../layouts/forms"
 import { AppSubmissionContext } from "../../../lib/AppSubmissionContext"
 import React, { useContext } from "react"
 
-export default () => {
+const ApplicationConfirmation = () => {
   const { application, listing } = useContext(AppSubmissionContext)
   const { initialStateLoaded, profile } = useContext(UserContext)
   const router = useRouter()
@@ -92,7 +92,7 @@ export default () => {
               <Button
                 styleType={AppearanceStyleType.primary}
                 onClick={() => {
-                  void router.push("/create-account").then(() => window.scrollTo(0, 0))
+                  void router.push("/create-account")
                 }}
               >
                 {t("account.createAccount")}
@@ -122,3 +122,5 @@ export default () => {
     </FormsLayout>
   )
 }
+
+export default ApplicationConfirmation