+++ sites/public/pages/applications/review/demographics.tsx
@@ -25,7 +25,7 @@ import {
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
 
-const Demographics = () => {
+const ApplicationDemographics = () => {
   const { conductor, application, listing } = useFormConductor("demographics")
   const currentPageSection = 5
 
@@ -76,7 +76,10 @@ const Demographics = () => {
       </FormCard>
 
       <FormCard>
-        <FormBackLink url={conductor.determinePreviousUrl()} />
+        <FormBackLink
+          url={conductor.determinePreviousUrl()}
+          onClick={() => conductor.setNavigatedBack(true)}
+        />
 
         <div className="form-card__lead border-b">
           <h2 className="form-card__title is-borderless">
@@ -171,4 +174,4 @@ const Demographics = () => {
   )
 }
 
-export default Demographics
+export default ApplicationDemographics