+++ ui-components/src/notifications/AlertBox.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react"
 import type { ReactNode } from "react"
-import { Icon } from "../icons/Icon"
+import { Icon, IconTypes, IconFillColors } from "../icons/Icon"
 import type { AlertTypes } from "./alertTypes"
 import { colorClasses } from "./alertTypes"
 import "./AlertBox.scss"
@@ -12,9 +12,11 @@ export interface AlertBoxProps {
   children: ReactNode
   inverted?: boolean
   className?: string
+  boundToLayoutWidth?: boolean
+  narrow?: boolean
 }
 
-const icons: { [k in AlertTypes]: string } = {
+const icons: { [k in AlertTypes]: IconTypes } = {
   alert: "warning",
   notice: "info",
   success: "check",
@@ -29,6 +31,8 @@ const AlertBox = (props: AlertBoxProps) => {
     colorClasses[props.type || "alert"],
     ...(props.inverted ? ["invert"] : []),
     ...(props.className ? [props.className] : []),
+    ...(props.boundToLayoutWidth ? [] : ["fullWidth"]),
+    ...(props.narrow ? ["narrow"] : []),
   ].join(" ")
 
   if (onClose) closeable = true
@@ -39,10 +43,14 @@ const AlertBox = (props: AlertBoxProps) => {
     }
   }
 
-  return showing ? (
-    <div className={classNames} role="alert">
+  let innerSection = (
+    <>
       <span className="alert-box__icon">
-        <Icon size="medium" symbol={icons[props.type || "alert"]} white={props.inverted} />
+        <Icon
+          size="medium"
+          symbol={icons[props.type || "alert"]}
+          fill={props.inverted ? IconFillColors.white : undefined}
+        />
       </span>
       <span className="alert-box__body">
         {typeof props.children === "string" ? <p>{props.children}</p> : props.children}
@@ -55,6 +63,15 @@ const AlertBox = (props: AlertBoxProps) => {
           &times;
         </button>
       )}
+    </>
+  )
+  if (props.boundToLayoutWidth) {
+    innerSection = <div className="alert-box_inner">{innerSection}</div>
+  }
+
+  return showing ? (
+    <div className={classNames} role="alert">
+      {innerSection}
     </div>
   ) : null
 }