+++ ui-components/src/text/Tag.stories.tsx
@@ -1,4 +1,8 @@
-import { AppearanceSizeType, AppearanceStyleType } from "../global/AppearanceTypes"
+import {
+  AppearanceSizeType,
+  AppearanceStyleType,
+  AppearanceShadeType,
+} from "../global/AppearanceTypes"
 import * as React from "react"
 import { Tag } from "./Tag"
 
@@ -25,8 +29,109 @@ export const pillSmall = () => (
   </Tag>
 )
 
-export const PillAndSuccess = () => (
+export const pillLarge = () => (
   <Tag pillStyle={true} styleType={AppearanceStyleType.success}>
-    Pill Success
+    Pill Large
   </Tag>
+)
+
+const Spacer = () => {
+  return <div style={{ height: "10px" }} />
+}
+export const pillColors = () => (
+  <>
+    <Tag pillStyle={true} styleType={AppearanceStyleType.primary}>
+      Primary
+    </Tag>
+    <Spacer />
+    <Tag pillStyle={true} styleType={AppearanceStyleType.secondary}>
+      Secondary
+    </Tag>
+    <Spacer />
+    <Tag pillStyle={true} styleType={AppearanceStyleType.success}>
+      Success Dark
+    </Tag>
+    <Spacer />
+    <Tag pillStyle={true} styleType={AppearanceStyleType.success} shade={AppearanceShadeType.light}>
+      Success Light
+    </Tag>
+    <Spacer />
+    <Tag
+      pillStyle={true}
+      styleType={AppearanceStyleType.accentCool}
+      shade={AppearanceShadeType.dark}
+    >
+      Accent Cool Dark
+    </Tag>
+    <Spacer />
+    <Tag
+      pillStyle={true}
+      styleType={AppearanceStyleType.accentCool}
+      shade={AppearanceShadeType.light}
+    >
+      Accent Cool Light
+    </Tag>
+    <Spacer />
+    <Tag
+      pillStyle={true}
+      styleType={AppearanceStyleType.accentWarm}
+      shade={AppearanceShadeType.dark}
+    >
+      Accent Warm Dark
+    </Tag>
+    <Spacer />
+    <Tag
+      pillStyle={true}
+      styleType={AppearanceStyleType.accentWarm}
+      shade={AppearanceShadeType.light}
+    >
+      Accent Warm Light
+    </Tag>
+    <Spacer />
+    <Tag pillStyle={true} styleType={AppearanceStyleType.closed}>
+      Closed
+    </Tag>
+  </>
+)
+
+export const pillListings = () => (
+  <div style={{ width: "200px" }}>
+    <Tag
+      pillStyle={true}
+      styleType={AppearanceStyleType.primary}
+      size={AppearanceSizeType.small}
+      fillContainer={true}
+    >
+      Draft
+    </Tag>
+    <Spacer />
+    <Tag
+      pillStyle={true}
+      styleType={AppearanceStyleType.success}
+      shade={AppearanceShadeType.light}
+      size={AppearanceSizeType.small}
+      fillContainer={true}
+    >
+      Scheduled
+    </Tag>
+    <Spacer />
+    <Tag
+      pillStyle={true}
+      styleType={AppearanceStyleType.closed}
+      size={AppearanceSizeType.small}
+      fillContainer={true}
+    >
+      Closed
+    </Tag>
+    <Spacer />
+    <Tag
+      pillStyle={true}
+      styleType={AppearanceStyleType.accentCool}
+      shade={AppearanceShadeType.dark}
+      size={AppearanceSizeType.small}
+      fillContainer={true}
+    >
+      Results
+    </Tag>
+  </div>
 )